fix(pr-comments): fix pr referenced comments
