+++ web-client/src/ducks/auth/facebook/types.ts
@@ -1,14 +1,8 @@
-import createActionTypeFactory from 'src/store/utils/createActionTypeFactory';
+import { asyncType } from '../types';
 
-const { asyncType: firebaseAsyncType } = createActionTypeFactory('FIREBASE');
-
-export const FIREBASE_FACEBOOK_LOGIN_POPUP = firebaseAsyncType(
-  'FACEBOOK_POPUP',
-);
-export const FIREBASE_LOGIN_REDIRECT = firebaseAsyncType('LOGIN_REDIRECT');
-export const TRIGGER_LOGIN_WITH_REDIRECT = firebaseAsyncType(
+export const FIREBASE_FACEBOOK_LOGIN_POPUP = asyncType('FACEBOOK_POPUP');
+export const FIREBASE_LOGIN_REDIRECT = asyncType('LOGIN_REDIRECT');
+export const TRIGGER_LOGIN_WITH_REDIRECT = asyncType(
   'TRIGGER_LOGIN_WITH_REDIRECT',
 );
-export const GET_LOGIN_REDIRECT_RESULT = firebaseAsyncType(
-  'GET_LOGIN_REDIRECT_RESULT',
-);
+export const GET_LOGIN_REDIRECT_RESULT = asyncType('GET_LOGIN_REDIRECT_RESULT');
